Needed restructure app clip
